Start complete reconstruction of postgresify.py
Severe problems found in "betänkande" documents, e.g. duplicate columns
in INSERT statements. The most reasonable way to correct this seems to be
a real parsing of the SQL (as opposed to just replacing with regexes).

This commit is the first where I import `sqlparse` - seems to work well,
but will require a little more work before we can import "betänkande" docs.

`sqlparse` installation is not added to Vagrant setup yet, so we have to
run  `pip install sqlparse` on the virtual machine to make it work.

To be continued...