Enable handling of some "betänkande" documents
postgresify now uses a mixture of regexes and SQL parsing. It seems that
we can successfully postgresify and execute the SQL in
http://data.riksdagen.se/dataset/dokument/bet-2010-2013.sql.zip with this code.

Added the url above to the set of documents to download and process in  init_db.sh.

Also added the sqlparsing dependency, so it should now be pip installed
with vagrant provision.

Hence, something like the following should work:
	~/app/riksdagen_sql$ bash clear_db.sh
	~/app/riksdagen_sql$ python get_data.py
	~/app/riksdagen_sql$ bash init_db.sh

It's not beautiful, but it works. To be continued...