Restructure yet again
View functions (in MVC language "controllers") are kept thin.
Their main function is to do routing and to parse request data.

At the moment, each .py file in app/views defines a blueprint. If the
number of views using a single blueprint becomes very large, it is
possible to break them into several files, perhaps as a package inside
views.

If some serious logic is needed to generate a page, let's put it
in "models". A stupid example in this commit shows how.